airflow/requirements.txt
# Core Airflow
apache-airflow==2.9.3
apache-airflow-providers-cncf-kubernetes==8.5.0

# Database
psycopg2-binary==2.9.9  # For PostgreSQL
SQLAlchemy==2.0.23

# gRPC
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.2

# Cloud Providers
apache-airflow-providers-google==12.3.0
apache-airflow-providers-amazon==10.2.0
apache-airflow-providers-microsoft-azure==8.2.0

# Integrations
apache-airflow-providers-redis==4.2.0
apache-airflow-providers-http==4.6.1
apache-airflow-providers-sqlite==3.5.0

# Monitoring
prometheus-client==0.20.0
sentry-sdk==1.40.0

# Development
pytest==7.4.4
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
mypy==1.8.0

# Utils
python-dotenv==1.0.0
python-json-logger==2.0.7
